worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-whisper>=20250625",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "redis>=7.1.0",
    "boto3>=1.34.0",
    "moviepy>=1.0.3",
    "ffmpeg-python>=0.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
known-first-party = ["worker"]